lab02/requirements.txt
absl-py==0.2.0
astor==0.6.2
bleach==1.5.0
gast==0.2.0
graphviz==0.8.2
grpcio==1.11.0
h5py==2.7.1
html5lib==0.9999999
Keras==2.1.5
Markdown==2.6.11
numpy==1.14.2
Pillow==5.1.0
protobuf==3.5.2.post1
pydot==1.2.4
pyparsing==2.2.0
PyYAML==3.12
scipy==1.0.1
six==1.11.0
tensorboard==1.7.0
tensorflow==1.7.0
termcolor==1.1.0
Werkzeug==0.14.1
